+++ src/Core/Dependency/CatalogDependencyResolver.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Th3Mouk\MaterializedView\Core\Dependency;
 
-use Doctrine\DBAL\Connection;
+use Th3Mouk\MaterializedView\Core\Database\Connection;
 use Th3Mouk\MaterializedView\Core\Definition\MaterializedViewName;
 use Th3Mouk\MaterializedView\Core\Registry\MaterializedViewRegistry;
 use Th3Mouk\MaterializedView\Core\Sql\ManagementMarker;